music 분리 game logger 오타 수정
